Desktop/Project 1Manufacturing - Predictive Maintenance/requirements.txt
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
shap
joblib
pyyaml
flask
matplotlib
